+++ server/index.js
@@ -1,9 +1,10 @@
-const fs     = require('fs'),
-      koa    = require('koa'),
-      serve  = require('koa-static'),
-      config = require('../config');
+import fs     from 'fs';
+import koa    from 'koa';
+import serve  from 'koa-static';
+import config from '../config';
 
-const app = koa();
+const paths = config.get('utils_paths');
+const app   = koa();
 
 // ------------------------------------
 // Response Time Header and Logging
@@ -14,16 +15,16 @@ app.use(require('./middleware/logger'));
 // ------------------------------------
 // Static File Middleware
 // ------------------------------------
-app.use(serve(config.inDist('client'), {
+app.use(serve(paths.dist('client'), {
   index : '__IGNORE_INDEX.HTML__'
 }));
 
 // ------------------------------------
 // View Rendering
 // ------------------------------------
-const template = fs.readFileSync(config.inDist('client/index.html'), 'utf-8')
-  .replace('<div id="mount"></div>', '<div id="mount">${content}</div>');
+const template = fs.readFileSync(paths.dist('client/index.html'), 'utf-8')
+  .replace('<div id="root"></div>', '<div id="root">${content}</div>');
 
 app.use(require('./middleware/render-route')(template));
 
-module.exports = app;
+export default app;