+++ server/middleware/logger.js
@@ -1,8 +1,8 @@
-module.exports = function *loggerMiddleware (next) {
+export default function *loggerMiddleware (next) {
   const start = new Date();
 
   yield next;
 
   const ms = new Date() - start;
   console.log('%s %s - %sms', this.method, this.url, new Date() - start);
-};
+}